Удалил папку с образцами функций
